Move cli to lib
